Hotfix: Lock quemu to a version that does not segfault (#55)
The build pipeline started failing recently. It was crashing with this error:
```
45.34 Preparing to unpack .../libc-bin_2.35-0ubuntu3.9_arm64.deb ...
45.35 Unpacking libc-bin (2.35-0ubuntu3.9) over (2.35-0ubuntu3.8) ...
45.60 Setting up libc-bin (2.35-0ubuntu3.9) ...
45.73 qemu: uncaught target signal 11 (Segmentation fault) - core dumped
46.18 Segmentation fault (core dumped)
46.24 qemu: uncaught target signal 11 (Segmentation fault) - core dumped
46.60 Segmentation fault (core dumped)
Sub-process /usr/bin/dpkg returned an error code (1)
```

I managed to dig a working version from the history of successful builds. If I managed to check correctlu that was using quemu 7.0.0 and the default version used (":latest") was using 9.2.0.

Doesn't seem like there's any tags any longer for the 7.0.0 images,